Make the script work for nested repeaters and flexible content fields with layouts
The nesting of repeaters and flexible field groups may have arbitrary
depth.

Make sure the addAcfDataToContent()-function can work with an object
with nested content fields.

I added a function deleteContent(), which deletes the data of a field
(repeater row or layout) when it is removed. This makes the function
updateRepeater() redundant.

The parameters updateYoast and updateRepeater of the setContent()
function are also redundant and have been removed. Therefore, the
function initContent() would only directly call setContent(), so it is
removed and replaced with a direct call to setContent().

Use jQuery's attr('data-<name>') instead of data('<name>'), because the
latter gives incorrect results for newly added rows/layouts.