Fix: Unable to remove unsupported field when inside a repeater or custom component
A JavaScript error occurred when removing an unsupported element,
because the ``parentContent`` was undefined. This problem only occurred
when the repeater or flexible content field had only unsupported
elements.
Solution: check if the ``parentContent`` is undefined before trying to
delete one of its properties.
Add this check also in the loop, for nested repeaters and/or flexible
content fields.
It is not possible to break out of a JavaScript ``forEach()``-function,
so use the jQuery ``.each()`` instead. Add support for the
``reverse()``-function on a jQuery array.